refactor rta reconnect organization
